Only show interwiki links dialog for named users
The 'Add interlanguage links' pop-up code was failing was a CORS
error when temporary users tried to click the 'Add interlanguage
links' link from the toolbar.

Previously, the pop-up had only been presented to logged-in users
and anonymous users were simply redirected by href to the edit page
on the repo. This determination was made by checking
`$user->isRegistered()`. Since temporary users are also "registered"
we replace this check in the `BeforePageDisplayHandler` with a call
to `$user->isNamed()` so that only named, logged-in users get the
`linkitem` javascript module loaded.

Bug: T370991
Change-Id: Idef3be24e0e22753a3c91680a890e1261e1d1388