eslint: Change wikimedia/client-es5 to client
Mainly motivated by seeing the disablement of es-x/no-object-assign in
the parent change – it makes no sense to me to have both rules that
require Object.assign() and rules that ban it in the same ruleset. Now
that MediaWiki requires ES6, I don’t think there’s any need to limit
ourselves to ES5 (and indeed, given that the parent change uses
Object.assign(), we do require ES6), so let’s just update the config.

This does mean that we need to disable the rules that want to force us
to immediately use let/const and arrow functions everywhere; someone™
can do that later instead.

Change-Id: I88d137b153bcda93e83f9580763d503272e04411