Do not set the event agent if it doesn't exist locally
It doesn't happen often, but on the repo side sitelinks can
be added by users who do not exist on the target wiki. Echo
cannot handle this (and it should be fixed in Echo as well).

Solve this by simply setting null agent which Echo allows.
I believe this is acceptable for the few events.
The notification puts emphasis on what happened rather than
who made it happen anyways.

Bug: T367638
Change-Id: I03b4367355dc5a3fc0c14aad5fdf19fbcd0caa3d