Don't add the now unused element in editing footer
An injected element shouldn't be needed as the linked change the core
will go using pseudo-element which isn't itself perfect but resolves
the two discovered regressions.

And the manual need of syncing with core and the implicit dependency
shows the situation here isn't ideal as the styling should be either
completely independent or having structured way for reuse from Core
changes however there is lacking infrastructure here as lack of Less
availability in extension-client as far as I know which makes me
think this is the best we can do for now.

However this implicit dependency isn't new actually as there was an
unnoticed breakage when mediawiki.icon use was removed from the
Core (not the module removal itself), this only makes it sync with the
changes in core but neither this nor my previous change doesn't
introduce nor removes the implicit dependency which should be
addressed sometime.

Depends-On: Id195ec4a20f11e12462bfc98362266a7b23dba19
Change-Id: Ib7aefc5fb18b127a8915eade5845f381ada31298