Pass language links as TitleValue, not as strings
This ensures that we don't reparse the title and accidentally convert
an interwiki prefix to a namespace.

Strictly speaking, this patch should not change Wikibase behavior, since
the string passed to ParserOutput::addLanguageLink() is not treated by
core as a "real" title but instead as a simple colon-delimited pair.
This patch however clarifies the behavior.

Bug: T204792
Bug: T363538
Change-Id: I0ecb8c2d138a4b8101859ee250649a6caff49d15