REST: Consolidate error status tests in EntityUpdaterTest
This combines 6 very similar tests into one with a data provider
providing the 6 different kinds of inputs and expected exceptions.

It also fixes a subtle issue in the tests by removing the use of
`expectExceptionObject` which only checks the type, but not the object's
contents. There was actually one test (the spam blacklist one) where the
expected exception did not match the one that was actually thrown.

Change-Id: I99e94db033086ff92f2e9b4148b1638041716dd5