Ensure ParserOutput::getTimeProfile() is valid for all ContentHandlers
ContentHandler::getParserOutput() calls
ParserOutput::resetParseStartTime() before invoking
ContentHandler::fillParserOutput(), and then calling
ParserOutput::recordTimeProfile() on the value that returns.  This
assumes that ::fillParserOutput() mutates the ParserOutput which it is
given, instead of creating a new ParserOutput.  If you create a new
ParserOutput, you are responsible for calling ::resetParseStartTime()
on it at an appropriate point.  The various content handlers in core
respect this (poorly documented) contract, but the wikibase
ContentHandlers did not.

Bug: T376433
Change-Id: I6bba35cea2e7213fd48e5f8d25e0543695b11263