Add link to revision in undo + restore summaries
In an entity's history summary, this makes the revision number
into a link. For the undo message, it links to the diff. For
the restore message, it links to the revision.

For consistency with `wikibase-entity-summary-undo`, the
`wikibase-entity-summary-restore` message is updated to include
a link to the revision author's talk page.

Bug: T194402
Change-Id: I185eea2ada806b786326375e41fdc94709b85512