Make phpunit dataProviders static, part 1
non-static dataProvider functions are deprecated as of PHPUnit 10.
This commit covers `repo/tests/phpunit/includes/Api/`, making
relatively small changes to the affected functions. In some cases,
they only needed to be marked as static. Others needed to change
the way they called other, already-static functions.

This commit the first of several in a series, broken up for easier
reviewing, and to minimize risks of changing too many tests
simultaneously.

Bug: T337154
Change-Id: Ief3e84e1034981a61419bcbeaa0cd816dbeed576