Make phpunit dataProviders static, part 2
non-static dataProvider functions are deprecated as of PHPUnit 10.
This commit makes relatively small changes to the affected functions.
In some cases, they only needed to be marked as static. Others needed
to change the way they called other, already-static functions, or ones
that didn't require much change to become static.

Bug: T337154
Change-Id: I7a1da05bbe7f9dfd840184ef742e41b226bc6d72