Make phpunit dataProviders static, part 3
non-static dataProvider functions are deprecated as of PHPUnit 10.
This commit covers `repo/tests/includes/ChangeOp` and makes relatively
non-complex changes to the affected functions.

In some cases, they only needed to be marked as static. Others needed
to change the way they called other, already-static functions, or ones
that didn't require much change to become static.

Making `getEntity` static in these abstract classes breaks tests in
`WikibaseMediaInfo`, due to a mismatch in implementation, and blocks
merging:
  `LabelsChangeOpDeserializationTester`
  `DescriptionsChangeOpDeserializationTester`
To facilitate the change, the affected tests are skipped in the depended
upon change, and will be re-enabled with the needed `getEntity` change
after this one goes through.

Bug: T337154
Depends-On: I55897282b853a47bd2a82910ac5ef4eab10f589e
Change-Id: I2b0873c7e54513f84011eaddc34d960b8490ceb1