Make phpunit dataProviders static, part 4
non-static dataProvider functions are deprecated as of PHPUnit 10.
This commit covers:
  `repo/tests/phpunit/includes/Hooks`
  `repo/tests/phpunit/includes/ParserOutput`
  `repo/tests/phpunit/includes/Parsers`
  `repo/tests/phpunit/includes/Rdf`

It makes relatively non-complex changes to the affected functions.
In some cases, they only needed to be marked as static. Others needed
to change the way they called other, already-static functions, or ones
that didn't require much change to become static.

Bug: T337154
Change-Id: Iaf6ea4535111303aa10b02234d09675177f60489