Make dataProviders static in tests with mocks
non-static dataProvider functions are deprecated as of PHPUnit 10.
This commit fixes another batch of dataProviders. Most of these
were creating mocks in the dataProvider functions, and relied on
an instance of the test class.

In some cases, setting up the mocks could be moved to the test
functions themselves. In other cases, the dataProvider now returns
a function to be called within the test function taking `$this` as
an argument.

Bug: T337154
Change-Id: I45e99808a89eb181b9e9186fd5b6eb6df78636dc