Document available wbformatvalue options
We’re about to declare other formatter options unsupported (change
Ifd1367eaff), so it stands to reason that the four validated options are
supported, part of the stable interface, and should be documented.

I first thought about putting this in docs/topics/, but I don’t think it
really fits there (and it would be confusing with the already existing
docs/topics/options.md). I think the parameter documentation works
pretty well (Special:ApiSandbox even makes the definition list
collapsable automatically), and there’s robust support for customizing
it and adding additional parameters (to avoid repeating the option names
– though they’re still part of the stable interface and can’t easily be
renamed, of course) both in ApiBase and in {{doc-apihelp-param}}.

Bug: T323778
Change-Id: I92931f3ddeef97ec22f89869c687f63f9fb596cc