Turn “data provider” into test
This method is named like a data provider, but looks like a test. As far
as I can tell, it’s been this way ever since it was introduced in change
I29c47c71a1 (commit 1e4fcd23f9). Amazingly, it works fine as a test and
did not break in the past ten years – we just need to rename it!

Bug: T337154
Change-Id: I4b4042c578d5232101879097933179895d768488