REST: Replace allOf workaround workaround in OAS
Iec2109e43045c4c43c0fa2571eef8730ad6275ad fixed an OpenAPI schema linter
rule violation by wrapping a single schema in an `allOf`. We don't fully
understand why this fixes the issue, and whether it's a JSON schema
feature or a quirk of the implementation we're using.

The problem itself appears to occur when declaring required properties
in the schema, while only listing some of those in the `properties`
section of the schema. An alternative fix, which we find less confusing
than the single element `allOf`, is to simply list all required
properties in `properties`.

Change-Id: I7beaf160c4fe9347412de05570c8902ba7bd4755