client: Make InjectRCRecordsJobTest data providers static
For provideMakeJobSpecification(), make getEntityChangeFactory() static
(doesn’t use $this anyway), and then use the fn ( self $self ) trick for
the getEntityChangeMock() call. (There’s already a $changeFactory
variable in the test, so rename the callable’s result to $expectedChange
and call the callable $expectedChangeFactory.) The $title was always the
same, so move that into testMakeJobSpecification().

provideConstruction() only provided one test case, so just inline that
into testConstruction() directly.

Bug: T337154
Change-Id: I921c388ae79877b2ef39d48b9d7b8c0877533b32