provideLuaData - remove override, make stub version static
The conditional in the WikibaseLibraryInProcessEntityCacheTest
implementation was always true when Scribunto is loaded, and would
call the parent method. When Scribunto wasn't loaded, it would
check for an implementation in a fake base class defined in Wikibase.

This commit removes the override, and adds a static `provideLuaData`
to the stub class, returning an empty array. Now, in all cases, the
test can rely on the parent method.

`provideLuaData` is still non-static in Scribunto's `LuaEngineTestBase`.
This likely means that Wikibase is not yet 100% ready for PHPUnit 10.
Making dataProviders static in Scribunto is covered by T337135

Bug: T380604
Change-Id: Iecf965ee1a8dfd1714d952541fdf6b57ad59b32b