Replace calls to deprecated `defaultOption` with `withDefaultOption`
The `data-values/interfaces` library deprecated the `defaultOption`
instance method on `FormatterOptions` in place of the method
`withDefaultOption` which leaves the original options un-mutated and
returns a copy of the options object with the default values set
(where not provided by the user).

Replace calls in Wikibase to `defaultOption` with `withDefaultOption`,
since we now have `data-values/interfaces` at 1.2.0.

Bug: T379481
Change-Id: I36e03dee15ae807f735dddf94871e6f775043a50