Fix bogus EntityAccessLimitException in RestrictedEntityLookup
This mistake was made in If4c2dd4 only a few days ago.

Note we are discussing quite a few possible behaviors here:

1. The moment the limit is reached everything starts to fail,
even querying entities that have been queried before. This is what
happened before If4c2dd4 and was not correct, I would argue.

2. Since If4c2dd4 there is only one failure when an entity is
accessed the first time, but not when it's accessed again. This
means a page can actually query as many entities as it wants
(even if the page is partially incomplete and there are error
messages on the page).

3. The moment the limit is reached only entities that have not been
queried before start to fail. This is what I propose here. Note how
entityHasBeenAccessed is used! The code actually makes this
assumption, but that was not correct.

Bug: T279069
Change-Id: I446a685ff7660a598875d4ed414520acc1cc042a