Add error handling for `continue` parameter to entityusage API
`wblistentityusage` and `wbentityusage` both accept a `continue`
parameter to filter larger responses. The continue parameter is
expected to be in the form `<pageId>|<entityId>|<aspect>`, but no
validation is in place to ensure that the parameter is in the
correct format before it is exploded into component parts.

Add validation and error handling to avoid unhandled user-facing
errors when an invalid `continue` parameter is supplied.

Bug: T322431
Change-Id: Ied42c1bace855a790b9ae79a3163c2f50bc249fa