Introduce TermsDomainDb
This introduces a new abstraction for the wikibase terms (labels,
descriptions, aliases) database, in order to enable having those tables
in a separate database.

For now, the TermsDomainDb is just a decorator around the existing
RepoDomainDb. In a next step, TermsDomainDb will be turned into an
interface with two implementations: one being that RepoDomainDb
decorator, and the other being a wrapper around a virtual domain
database connection. The TermsDomainDbFactory will then decide which
one to create based on the virtual domain config.

Bug: T351802
Change-Id: I99866a3c4ddec8585b187468a5224d5ce1022736