repo/includes: Fix implicit nullable types
This leaves repo/rest-api and repo/tests still to be done.

In ContentHandlerEntityTitleLookup, I made the InterwikiLookup fully
required, because as far as I can tell it’s never null in any of the
existing callers. (It was apparently split out of a different class –
maybe *that* class’ InterwikiLookup could sometimes be null, I haven’t
looked into it.)

Bug: T379509
Change-Id: I77f042786ae38458941b60d6d1398ed5ff262e4f