client: Turn invalid continue warning into error
The previous warning message wasn’t defined anywhere, resulting in the
warning ⧼apiwarn-ignoring-invalid-continue-parameter⧽ being shown. I
also think an error makes more sense anyways: if the continue parameter
in invalid, we can’t know much about what the user intended with the
request, but it seems reasonable to assume that they won’t be interested
in results with no continuation (from the beginning).

In the task description, I pointed to ListSubscribers as an example we
could emulate, but that uses $this->errorReporter->dieError(), which
turns out to be a repo-only class. (And one we should probably
eventually get rid of entirely, per T304945.) Fortunately, it turns out
MediaWiki core has a dedicated method just to report “bad continuation
parameter” errors, so let’s just use that.

Bug: T322431
Change-Id: Ief02a4f0a780897d572ad926c74a042f92d703fa