Add config option to fix s:, ref:, v: namespace prefix
These should have used the node prefix (yielding wds:, wdref:, wdv: on
Wikidata), but accidentally used the predicate prefix instead. Introduce
a config option to fix that; as this changes the RDF output (remains to
be decided whether that’s considered a significant or breaking change,
but it certainly needs an announcement), put it behind a config option.

Bug: T384344
Change-Id: I5e9da89e5096146898f9ac087a75942990868d5e
Depends-On: I458d67dcfb0a755f31896cb8af6489fd97aeb731