Handle missing `monthonly` format in MwTimeIsoFormatter
I9aefa02a8355c9cca6ac750b2bf82a7734338640 introduced the case that
a null string might be passed to `preg_match` in `getMonthFormat`.
This worked fine until PHP 8.1, which deprecates `preg_match` on
null strings.

Check that the format is defined before running the `preg_match` in
`getMonthFormat` for the case that no `monthonly` format is defined
and otherwise fallback to the default `F`.

Bug: T384867
Change-Id: Ib7392d0bc20670e7b2c5189caefb4f0379603ac9