Handle null date format in MwDateFormatParserFactory
MediaWiki can sometimes return a null date format, such as when the
'monthonly' date format type and 'default' date format preference are
used in certain languages; avoid setting OPT_DATE_FORMAT in that case,
so that DateFormatParser falls back to its hard-coded default instead.

Unfortunately, I don’t see a great way to test this at the moment.
MwDateFormatParserFactoryTest can be made to generate test cases where
the date format ends up as null by adding something like

    'default' => TimeValue::PRECISION_MINUTE,

to the $dateFormatPreferences in validInputProvider(), but then we
already get a PHP warning (and thus a test failure) as soon as we call
$language->sprintfDate( $dateFormat, $mwTimestamp ) immediately after
$language->getDateFormatString( $dateFormatType, $dateFormatPreference )
– so the test never even reaches our own code.

Bug: T384963
Change-Id: I07645ab4b0f27dcf842070d64d93b69da9f7da04