Rename the `tmpEnableMulLanguageCode` flag to `enableMulLanguageCode`
As part of the cleanup after the rollout of the `mul` feature, make
the feature flag permanent and rename it to `enableMulLanguageCode`.

The dependent config change should be pushed out first so that both
the old and new value are enabled for the production config before
support for the old value is removed from Wikibase.

Depends-On: Ifac9c6761c16aac244d96cca60f811e1fbcd05ed
Bug: T330217
Change-Id: I48d4c410e30c0fc9a631bc851c21603d773bda19