Add feature flag for ScopedTypeaheadSearch
This temporary feature flags controls whether the ScopedTypeaheadSearch
feature is enabled or not. As the feature isn’t implemented yet (a
prototype is in progress, see T384730), for now all this actually
“enables” is replacing the search bar with a message (not translatable,
because this is only meant for developers anyway). You do not want to
enable this setting yet. (We also don’t enable it in CI yet.)

Because every ResourceLoader module definition carries a small cost,
even if the module is never loaded (compare T203696 and [1]), do
something slightly unusual and register the module in RepoHooks, rather
than Resources.php, only if the flag is enabled, so that the module
isn’t even defined in production yet. At some point it should move into
Resources.php after all. (IIUC, ScopedTypeaheadSearch will effectively
replace current functionality of the wikibase.vector.searchClient
module, so maybe we can replace that module and thus keep the total
module count the same.)

[1]: https://phabricator.wikimedia.org/project/profile/4268/

Bug: T384735
Change-Id: I7d8b53f9a8ae925d5a9134b33581628d31e0f220