WikiPageUpdaterTest: Use StatsFactory::newUnitTestingHelper()
This makes for a relatively nice interface once you figure out how to
use it properly (removing the flush() calls is important; when testing a
count metric, you want sum(), not count()), at least for the new
statslib metrics. It offers no support at all for testing the copying to
statsd, so reuse the old test strategy for that (inject a mock with
expected calls). The resulting split of the expected metrics – statsd
near the beginning of the test and statslib near the end – is annoying,
but will be removed once statsd support is gone.

Bug: T359248
Bug: T368740
Change-Id: I41204bdfcebb60503edf0b9eb6f938d4cdeb4b77