Add config option to make somevalue hashes use URI
Currently, the hashes are based on the namespace name, which we want to
change soon (from v: to wdv:). Since that will break the hashes, let’s
take this opportunity to make the hashes semantically cleaner by basing
them on the URI instead (e.g. <http://www.wikidata.org/value/>). This
change adds a feature flag for that, without turning it on anywhere yet.

Also add a somevalue test case to SnakRdfBuilderTest, which was
completely missing so far. (The expected hash is currently the “old”
hash – it’ll be updated along with other tests in change If66c288bca.)

To simplify the dependency injection for this temporary flag, make it a
public member of RdfVocabulary, where SnakRdfBuilder can easily access
it without needing it injected separately.

Bug: T384344
Change-Id: I5a1648853bed50b3e9ff7b8010324d314034f787