REST: Delete test user credentials if login fails
If the test user credentials are no longer valid, e.g. due to the dev
wiki being reinstalled or the user being deleted, this helper function
now discards the credentials file automatically. It still errors so that
this action remains visible and doesn't cause us to silently create new
users when we expect them to be reused.

Change-Id: Id3f18028c27057a75b76a9720b474cacef3a6394