REST: Fix flaky mocha tests caused by AuthTest.js
Occasionally, mocha tests in the `mediawiki-quibble-apitests-vendor-php74` CI job fail with the following error:
```
{
  code: 'permission-denied',
  message: 'Access to resource is denied',
  context: { denial_reason: 'resource-protected' }
}
```

I suspect this is due to tests in `AuthTest.js` protecting an item/property that is also used in other tests. `AuthTest.js` uses the `describeWithTestData()` helper function which in turn uses the `getItemId()` helper function which enable the reuse of a single Item across tests.

Adding a `createNewItem` param to `describeWithTestData()` allows `AuthTest.js` to request that a new Item is created for its tests.

Change-Id: I0ba157a6bffdf8facc9a99905ae9287e71f8d4e1