Fix REST API client generator
The GitHub action failed due to being unable to find the openapi.json
file, which moved recently and the path wasn't updated correctly.

This patch also removes a workflow step that copied the openapi.json
file into the current working directory, which turned out not to be
necessary, because the repo as a whole gets mounted into the container
generating the client.

Bug: T386385
Change-Id: If7a0a7c6410cf9de9dcd6be8dae27d37343ce7a4