ScopedTypeaheadSearch searches local wikibase
For ease of development, ScopedTypeaheadSearch was hardcoded to
search wikidata.org. This commit removes the hardcoded references
to wikidata.org and replaces them with locally-relevant values.
This includes using the api for typeahead searching, pointing the
form action at a search url, and using the name of the site for
the input placeholder and in the footer text.

Bug: T387481
Change-Id: I0cce1721454fd2ae74892ce6fe4eee1881ba208e