Add hook to allow extensions to register scoped typeahead search scopes
Since T385791 the list of scopes for the scoped typeahead search is
determined by dynamic configuration based on which entity namespaces
are enabled for search. We want also to be able to search for Entity
Schemas, but EntitySchema is not an entity type.

Add a `WikibaseRepoSearchableEntityScopes` hook to allow extensions to
register scopes for the scoped typeahead search without needing to
register complete entities.

Bug: T387986
Change-Id: I10d4c8d9838a94bbdc3f35cc8961f15f957d6b54