Replace RevisionRecord::getPageAsLinkTarget with getPage
This makes more sense, in my opinion. A revision in this context is
not something on a remote wiki we can only link to (via LinkTarget).
We actually have the page and can use it.

This came up while working on Iecd457a. It's closely related but
doesn't block said patch.

Change-Id: Iffc428465d475c1947998e16394d3a8cebf12650