Set up Jest tests for ScopedTypeaheadSearch
Just the initial setup, partially based on [[mw:Vue.js/Testing]] [1],
though our Jest version is newer and consequently we need @vue/vue3-jest
rather than vue-jest.

We need to mock Codex because the ResourceLoader-generated magic file
with just the components specified in the module definition doesn’t
exist in a test context. We can just pass in the full Codex. (This does
mean that we risk not updating the RL codexComponents when we need an
extra component – the tests wouldn’t catch this.) Installing Codex via
npm is a bit ugly – it means this version will have to be bumped with
each Codex release – but I’m hesitant to pull it in from MediaWiki core
instead (something like ../../../../../resources/lib/codex/codex.cjs)
because that would mean the tests won’t work inside fresh-node (unless
you run fresh-node in the MediaWiki core directory rather than the
Wikibase directory).

I wasn’t able to find a way to mock scopedTypeaheadSearchConfig.json via
jest.mock(), so use a real file for that, at least for the time being.

[1]: https://www.mediawiki.org/wiki/Vue.js/Testing

Bug: T385790
Change-Id: I60706fc00f18adf1fa5e2b7baa4c005cd8e742f4