Use configured namespaceAliases for scope selection
The existing search functionality allows a user to specify a
namespace by prefixing the search term with a namespace alias
and a colon. This should continue to work in Scoped Typeahead
Search. In a previous iteration, this was accomplished via a
hardcoded mapping of one-letter prefixes to their scopes.

This commit uses namespace aliases, which are configured by
`$wgNamespaceAliases`, and accessible via `mw.config` instead.
Since these are not limited to a single letter and `:` character,
the code is also updated to be flexible with the prefix length.

Bug: T386201
Change-Id: I57411e4c473b5e734696005532ff7271cbb1e750