Use @color-disabled for references heading
Note that this color is only visible when JavaScript is disabled. (Back
in change I29a0c1f376 / commit e3837b0f7c, it was also visible on old
revisions and diffs, but this was fixed in the meantime, probably in
change Id05bfcc219 / commit 16002942a5.) Normally, the reference
togglers are styled blue like links (cf. T389345).

Bug: T389343
Change-Id: Ieb184e8a237be9ca4a1c25f2ae20f58b01acd279