Fix layout and colouring issues with Scoped Typeahead search
Adjust the scoped typeahead search so that the search dropdown is
only as wide as the text input (as opposed to as wide as the text
input plus scope dropdown).

Fix the alignment of the search icon at the bottom of the search
dropdown to match the left margin of the items in the search
dropdown.

Change the border colour of the component to `@border-color-base`,
and change the character spacing inside the search box so that the
texts is more compact.

Bug: T389228
Change-Id: Ie744d28a8f52f249ce05a8b6133085aaab825948