Replace option assignments that use $.extend with Object.assign
The `no-jquery/no-extend` eslint rule recommends against the use of
jQuery's `$.extend` construct. Since `$.extend` does a lot of
different things, it can be hard to reason about what is happening
in the code.

Where `$.extend` has been used in the test code in ways that are
equivalent to `Object.assign`, replace the jQuery implementation
with the Javascript-native version.

Bug: T379354
Change-Id: I5b3b78ff95d0198b85b2fe3e9a669ba3e3a42a68