Re-enable the no-jquery/no-extend eslint rule
Almost all usages of `$.extend` in the Wikibase code have been
removed and the eslint `no-jquery/no-extend` rule has been enabled.

There are two remaining usages, which have been covered with
eslint-disable-next-line no-jquery/no-extend, which concern the
implementation of Snak view variations.

These last instances will be addressed in T390181, but in the meantime
we can re-enable the rule so that no new usages of `$.extend` end
up in the codebase.

Bug: T379354
Depends-On: I3a3e6c8d1400be493bda0dbeac9b6f4941f3589b
Depends-On: Ic0589c10ae6d89d0ef241a5981c4cb12ca516ad8
Depends-On: I5b3b78ff95d0198b85b2fe3e9a669ba3e3a42a68
Change-Id: I1cf8bfd44c0a610c18796df77e32d0c2456254ff