Override some jQuery UI colors for dark mode
jQuery UI sets some hard-coded colors (not Codex tokens) for elements
that end up being inherited by our elements, such as text inputs when
editing references (because the reference view is ui-widget-content; the
main snak and qualifiers are not affected). It seems unlikely that this
will be fixed in the jQuery UI styles that Vector ships, so override the
colors in our own styles. (Limit the overrides to .wikibase-entityview
just in case someone else is using jQuery UI elsewhere on the page where
they need the original light-mode colors.) @color-base (#202122) isn’t
the same as the previous color (#362b36), but it’s a relatively close
match, the best semantic fit, and the token that UX specified.

Bug: T389633
Change-Id: I987aa250c8f7de10322daa20fb806dc7e81d6624