Add Item and CustomItem classes as properties to `$.ui.ooMenu`
In I3a3e6c8d1400be493bda0dbeac9b6f4941f3589b we removed confusing
usages of `$.extend`, which both copies properties between objects
and extends class prototypes. In porting the Item and CustomItem
class definitions, the patch saved the new class definitions to
`$.ui.ooMenu`'s prototype, when in fact it is a global object and
the class definitions need to be added as properties.

Add the class definitions as properties so that the classes can
be extended.

Bug: T390949
Change-Id: I33e39fd75ba0f7ddc9aa9d1999147807fe79635d