Refactor snakview variations to remove usages of `$.extend`
The snakview Variations (NoValue, SomeValue, Value) use jQuery's
`$.extend` and the `util.inherit` function to dynamically create
classes at runtime. This avoids a very small amount of boilerplate
at the expense of using the `$.extend` construct that we would like
to remove from our code.

Since this code is not considered stable, and since we expect that
the only callers of `variations.variation()` are the three variation
classes in the Wikibase codebase, we can simplify the code and
remove some of the more exotic capabilities of the snakview extension
model.

Refactor the snakview Variations to use Javascript-native classes
and class inheritance. Remove the remaining references to `$.extend`.

Bug: T390181
Change-Id: Ie63ebd37282970eb37531ac30a500973039a6728