Rework DatabaseMatchingTermsLookup
Implementing the OFFSET $options as expected is more easily achievable if DatabaseMatchingTermsLookup only executes a single SQL query.

Uses of the MatchingTermsLookup interface only ever passes a single $entityType to getMatchingTerms(). Removing the option for $entityType to be an array allows us to simplify the DatabaseMatchingTermsLookup implementation and effectively implement the OFFSET $options.

Bug: T390693
Change-Id: Ib217d1c94f4f0bf41508d967a518777629be54bd