GitHub secondary CI: Bump actions/cache
The previous version broke for unknown reasons:

>  Error: Unable to resolve action actions/cache@v3.4.0, package version
> not found

Let’s just update to the latest version (as of this writing) and hope
that works better.

In a curious interpretation of semantic versioning, v4.0.0 [1] has no
significant changes to its predecessor (the major version bump appears
to be due to updating the runtime from node16 to node20, which need not
concern us), whereas v4.2.0 [2] and v3.4.0 [3] brought important
changes.

[1]: https://github.com/actions/cache/releases/tag/v4.0.0
[2]: https://github.com/actions/cache/releases/tag/v4.2.0
[3]: https://github.com/actions/cache/releases/tag/v3.4.0

Bug: T391455
Change-Id: Ided04e250c24ff845753c08cb0824a9f5d19e6b7