Update reference tab bar header for dark mode
Override the background so that it matches the rest of the statement
view, making it look “transparent”. Also remove the rounded corners,
which aren’t visible in light mode but look ugly. In jQuery UI 1.12 [1]
or later, we could do this by specifying the `classes` option when we
call .tabs() in referenceview.js _createTabs(), but we’re still on
jQuery UI 1.9 [2], so do it in CSS instead. (Note that we already do the
same thing a few lines above, under “Clear default tabs ui styling”.)

[1]: https://api.jqueryui.com/1.12/tabs/#option-classes
[2]: https://api.jqueryui.com/1.9/tabs/

Bug: T389633
Change-Id: If06d1bc52dda9ca18c92ab41b9adfc5d485b5b51