Migrate GetEntityContentModelForTitle hook
Add it to the hook runner, inject that into ContentHandlerEntityIdLookup
and call it there, and introduce a hook interface that can be
implemented by hook runners and handlers. The documentation is adopted
from docs/topics/hooks-php.md (with “content type” changed to “content
model”, because “content model” is the MediaWiki concept while “content
type” refers to HTTP / MIME).

Bug: T391444
Change-Id: If694f071879840bd8c2158b7fbbca4ab8b57b1e7