Add WikibaseRepoHookRunner to service container
I think adding the hook runner there makes sense, and it seems to be
done in some other extensions. ScopedTypeaheadSearchConfig can then
directly get the hook runner injected instead of creating it from the
injected hook container; on the other hand, ScopedTypeaheadCodexModule
gets its hook container injected differently (by ResourceLoader), so we
continue to create the hook runner directly there.

Also add a test for the hook runner based on HookRunnerTestBase. This
ensures that the hook runner continues to follow the relevant patterns
as we update it in follow-up changes.

Bug: T391444
Change-Id: I6280f14407721d2ef3a8d680d3be856252c0fd5b