Migrate WikibaseChangeNotification hook
Remove the ability to inject a custom hook name into
HookChangeTransmitter, which was unused except in the test.

Also clean up phpdoc references to ChangeNotificationChannel, which
never existed in the main Wikibase.git history; this is the old name of
ChangeTransmitter, renamed in PS4 of change If2466a07f6 (commit
5658038979) which introduced it (but left some occurrences of the old
name behind).

Bug: T391444
Change-Id: I6540539211bff73196d73fe98435331bd6f9c8d5