Clean up hooks documentation
Move the remaining documentation to the hook interface and update those
interfaces in general (these two hook interfaces predate the current
batch of work, that’s why they were inconsistent); add a full stop at
the end of all the “See X” sentences; and remove Markdown link
references that became unused but we forgot to remove before.

Note that the change to onWikibaseHandleChange(), adding the
EntityChange parameter type, is compatible – classes implementing the
interface are allowed to widen the parameter type to (implied) mixed.
(Though there are no known implementations of this hook anyway – the
only known handler is EchoNotificationsHandlers::onWikibaseHandleChange,
which is still a static method.)

Bug: T391442
Bug: T391444
Change-Id: Ida3b3a5b266cb43a176b5418738fbad8a44c238d