Fix abortability of two repo hooks
WikibaseRepoSearchableEntityScopesHook was recently introduced in change
I10d4c8d983 (commit 21154c39a0); its hook interface returns void, and we
never checked the HookContainer::run() return value, but we neglected to
run it as unabortable. Fix that.

WikibaseRepoOnParserOutputUpdaterConstructionHook was recently migrated
in change Icbb9f96e18 (commit f3561b8be4), and while I marked the hook
runner method as returning void, I neglected to do so for the interface
itself. Given that the interface is so new, and has no known
implementations yet, I think we can just fix it even though this is
technically a breaking change. (Not that Wikibase has a stable interface
policy for its PHP code anyways.)

Both of these mistakes are detected by an upcoming improvement to
HookRunnerTestBase in MediaWiki core.

Bug: T391444
Change-Id: If37b72896f18d751bfaa19e6d340881319587b5f
Needed-By: I858d5dc02c65a0c68081214c4c71a7d816c0a524