Remove skipped test of event behaviour in scoped typeahead search
The ScopedTypeaheadSearch spec included a skipped test of the
event emitting behaviour, which would have asserted than a model
update event is emitted when the select's value is changed.

Looking at the docs [1] and a github issue about two-way event
binding [2], my conclusion is that direct updates to model values
via `ref`s do not necessarily trigger an update event.

The preceeding test (`when the input matches a configured prefix`)
makes an assertion that `api.get` is called, which assures us that
the `watch` on `selection` actually gets triggered. Making
assertions about the exact event behaviour seems to be unnecessary
to assure the correct behaviour of the component.

1. https://vuejs.org/guide/components/v-model.html
2. https://github.com/vuejs/core/issues/10729

Bug: T391841
Change-Id: I716ac856f12351efea906c9bda1560c391488b88