Move GetEntityContentModelForTitleHook to data-access
When I moved this hook in change If694f07187 (commit ec29535cf5), I
failed to notice that it’s used not only in repo, but also data-access
(and, via that, in client). Move the hook interface to data-access (it’s
only implemented in one other extension, where we can temporarily remove
the implementation, see Depends-On), introduce a hook runner for
data-access, and use the hook interface there.

While we’re touching the hook, also update its documentation a bit – the
previous version (copied from hooks-php.md) was somewhat outdated, as
the call to the hook had actually moved to a different class in change
I99826cecc2 (commit f7e2f9ee9d) four years ago. I don’t think including
this information is very useful (the call to the hook isn’t hard to
find), so just remove it.

Bug: T391671
Change-Id: I3de74b320be45e6c2464a2dddb9de4490295333d
Depends-On: I4aef41c262921bd2216eef1e7651be0c1ed4d876