Activate the scoped typeahead search when the search box is focussed
Because we are not able as an extension to make fundamental changes
to Vector's search box HTML, we can only add our Wikibase scoped
typeahead search functionality using Javascript after page load. To
support a smoother user experience, we decided to animate the
scoped typeahead search with a CSS transition from the search that
the Vector skin provides by default.

I83287b0da588 introduces a hook that allows us to replace the wiki
search module - register for that hook and abort it so that
subsequent subscribers (including the skin) do not register their
own search module.

Bug: T392109
Change-Id: I7be7f5794a864d9a12c6527323d159e18fdcb6a3